Move gtm-related type to another declaration file
