+++ react/typings/events.d.ts
@@ -147,16 +147,6 @@ export interface CartData extends EventData {
   orderForm: OrderForm
 }
 
-export interface AnalyticsEcommerceProduct {
-  id: string
-  name: string
-  category: string
-  brand: string
-  variant: string
-  price: number
-  quantity: number
-}
-
 interface CartItemAdditionalInfo {
   brandName: string
   brandId: string
