+++ react/typings/gtm.d.ts
@@ -1,0 +1,9 @@
+export interface AnalyticsEcommerceProduct {
+  id: string
+  name: string
+  category: string
+  brand: string
+  variant: string
+  price: number
+  quantity: number
+}
